core: Creates four functions:
- fold that receives an array, a function and an accumulator, in this order,
and applies the function in the elements of the array starting on the left.

- foldRight that receives an array, a function and an accumulator, in this order,
and applies the function in the elements of the array starting on the right.

- reduce that works just like the method [].reduce when you don't specify an accumulator.
The arguments should be an array and a function.
The starting value of your accumulator must be the first value of the array.
If your array is less than 1 argument you should throw an error.

- reduceRight like reduce, from the last value to the first